Have `type-case` syntax errors point to the right location.
The implementation of `type-case` has separate code for transformation
and for error reporting. I forgot to update the latter when allowing
multiple expressions in clause bodies.

For correct uses of `type-case`, everything worked fine. But for some
erroneous uses of `type-case`, where an error (e.g., an empty clause)
was in a clause that was after a multi-expression clause, the error
reporting would (mistakenly) flag the multi-expression clause as the
error. That would lead to unrelated, now ok clauses being flagged as
erroneous.

Note: `syntax-parse`'s combining of transformation and error checking
would have probably prevented that mistake.

Thanks to Arthur Mergner for providing an example.